Moved test_apple_xlocation() to test_icalendar.py
