+++ src/icalendar/tests/test_icalendar.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-import textwrap
+import icalendar
+import os
 
 from icalendar.tests import unittest
 
@@ -281,3 +282,17 @@ class IcalendarTestCase (unittest.TestCase):
         from ..parser import q_join
         self.assertEqual(q_join(['Max', 'Moller', 'Rasmussen, Max']),
                          'Max,Moller,"Rasmussen, Max"')
+
+
+class TestEncoding(unittest.TestCase):
+
+    def test_apple_xlocation(self):
+        """
+        Test if we support base64 encoded binary data in parameter values.
+        """
+        directory = os.path.dirname(__file__)
+        with open(os.path.join(directory, 'x_location.ics'), 'rb') as fp:
+            data = fp.read()
+        cal = icalendar.Calendar.from_ical(data)
+        for event in cal.walk('vevent'):
+            self.assertEqual(len(event.errors), 0, 'Got too many errors')
